Error Types

PowerDynamics.PowerDynamicsError — Type

Abstract super type of all PowerDynamics.jl Errors.

PowerDynamics.NodeDynamicsError — Type

Error to be thrown if something goes wrong during the node dynamics construction.

PowerDynamics.StateError — Type

Error to be thrown if something goes wrong when creating or modifying states.

PowerDynamics.GridSolutionError — Type

Error to be thrown if something goes wrong during power grid solving

PowerDynamics.OperationPointError — Type

Error to be thrown if something goes wrong during the operation point search.