PowerDynBase.jl

PowerDynBase.jl

PowerDynBase.AbstractAlgebraicGridDynamics — Type.

Abstract super type for all grid dynamics represented by DAEs.

PowerDynBase.AbstractAlgebraicNodeDynamics — Type.

Abstract super type for all node dynamics represented by DAEs.

PowerDynBase.AbstractDAEVariable — Type.

Abstract super type for all Variables for DAE-type node dynamics.

PowerDynBase.AbstractDEVariable — Type.

Abstract super type for all variables that AbstractNodeDynamics sub types can be called with.

PowerDynBase.AbstractNetworkFunction — Type.

abstract type AbstractNetworkFunction{T<:AbstractNodeDynamics, M<:AbstractMatrix} end

Abstract super type of all functions that define how a differential equation for the whole network / power grid behaves, e.g. the full right-hand-side function of the ODE.

PowerDynBase.AbstractNodeDynamics — Type.

Abstract super type for all abstract node dynamics types.

PowerDynBase.AbstractODEVariable — Type.

Abstract super type for all Variables for ODE-type node dynamics.

PowerDynBase.AbstractOrdinaryGridDynamics — Type.

Abstract super type for all grid dynamics represented by ODEs.

PowerDynBase.AbstractOrdinaryNodeDynamics — Type.

Abstract super type for all node dynamics represented by ODEs.

PowerDynBase.AlgebraicGridDynamics — Type.

TBD

source

PowerDynBase.AlgebraicNodeDynamics — Type.

DOCS TBD!

source

PowerDynBase.BaseState — Type.

    BaseState(grid, vec)

Encode a state vector and the corresponding rhs information.

Keyword Arguments

grid is a GridDynamics instance that contains the overall system rhs.

vec is a state vector of the system who's length is given by the total number of internal and voltage variables.

Indexing

In an instance b of of a BaseState behaves like an Array, i.e. you can access the $j$-th element of the state vector (and set it to a value $ξ$) by calling b[j] ( = ξ ).

PowerDynBase.DAEVariable — Type.

Variables for DAE-type node dynamics.

source

PowerDynBase.ODEVariable — Type.

Variables for ODE-type node dynamics.

source

PowerDynBase.OrdinaryGridDynamics — Type.

TBD

source

PowerDynBase.OrdinaryGridDynamicsWithMass — Type.

TBD

source

Base.view — Method.

Extend view from arrays to subtypes of AbstractDEVariable.

PowerDynBase.DynamicNode — Method.

See DPSABase.@DynamicNode.

PowerDynBase._GridDynamics — Method.

Create for each subtype of DPSABase.AbstractNodeDynamics the corresponding subtype of DPSABase.GridDynamics.

PowerDynBase.checkLY — Method.

Check whether the admittance laplacian has no purely nodal admittances, i.e. that the sum of columns and rows equals to zero.

PowerDynBase.complexview — Method.

Extend complexview from arrays to subtypes of AbstractDEVariable.

PowerDynBase.complexview — Method.

Interpret (part of) an array of real values as an array with complex values.

PowerDynBase.create_DEVariable — Method.

Basically, this macro generates all the constructors (internal and external) for a subtype of AbstracDEVariable.

If you really want to understand this macro, uncomment the println(ex) statement at the end and check the resulting generated expression.

PowerDynBase.excomparison — Method.

Create an expresseion where == is applied between all the expressions given as argument here.

PowerDynBase.getDEVariableType — Method.

Identify each subtype of AbstractNodeDynamics with its corresponding subtype of AbstractDEVariable

PowerDynBase.internal_unitranges — Method.

Get the unit ranges that indicate where in the array the internal variables for each of the nodes is saved.

PowerDynBase.mapfields — Method.

function mapfields(f, s, args...)

Applies f to all fields of (the struct) s giving args... as additional arguments.

PowerDynBase.nint — Method.

Get number of internal arguments of the node.

PowerDynBase.nodeiterator — Method.

nodeiterator(rhs::NetworkRHS, x::AbstractDEVariable, t)

Distribute the values in x over all the nodes that are summarized in rhs.

PowerDynBase.rhs2root — Method.

A function converting a rhs-type function to a root-type function.

PowerDynBase.total_nint — Method.

Get the total number of internal variables for an array of node dynamics.

PowerDynBase.total_nvars — Method.

Get the total number of dynamic variables for an array of node dynamics.

This is basically the (real) dimension of the system, hence the sum of internal dynamic variables + 2*(number of nodes = number of complex voltages). The 2 is due to the fact that the complex voltages are treated as two real variables.