Node Dynamics Types

Node Dynamics Types

In this section, the implemented general types of node dynamics with the corresponding helper functions and constants are introduced. The documentation is done for each type below. The main types are:

PowerDynBase.OrdinaryNodeDynamics

PowerDynBase.OrdinaryNodeDynamicsWithMass

PowerDynBase.OrdinaryNodeDynamics — Type.

OrdinaryNodeDynamics(;rhs, n_int)

The type representing the dynamics of a node that is described via ODEs.

Each node $a$ has the complex voltage $u$ and $n$ real internal variables $y_1, \dots, y_n$, so it generally describes a system of ordinary differential equation as

\[\frac{du_a}{dt} = f_u(u_a, {i_c}_a, y_1, \dots, y_n) \\ \frac{dy_{ak}}{dt} = f_k(u_a, {i_c}_a, y_1, \dots, y_n)\quad \forall k = 1, \dots, n.\]

$f$ is represented by rhs field of OrdinaryNodeDynamics.

the general signature of rhs is

rhs(dint_dt::AbstractVector,
    u::Complex,
    i::Complex,
    int::AbstractVector,
    t,
    )::Complex

Input

u is the complex voltage $u$

i is the complex current $i$

int is the array of internal variables $y_1, \dots, y_n$

t is the time $t$

Output

the (complex) return value describes $\frac{du}{dt}$

rhs writes values in dint_dt describing the left-hand side $\frac{dy_1}{dt}, \dots, \frac{dy_n}{dt}$

source

PowerDynBase.OrdinaryNodeDynamicsWithMass — Type.

OrdinaryNodeDynamicsWithMass(;rhs, n_int, m_u, m_int)

The type representing the dynamics of a node that is described via ODEs.

Each node $a$ has the complex voltage $u$ and $n$ (= n_int) real internal variables $y_1, \dots, y_n$, so it generally describes a system of ordinary differential equation with a voltage mass $m_u$ and internal masses $m^{int}_1, \dots, m^{int}_n$ as

\[m_u\frac{du_a}{dt} = f_u(u_a, {i_c}_a, y_1, \dots, y_n) \\ m^{int}_k\frac{dy_{ak}}{dt} = f_k(u_a, {i_c}_a, y_1, \dots, y_n)\quad \forall k = 1, \dots, n.\]

As we assume that all masses are binary (either 1, or 0), that means, one can implement semi-explicit differential algebraic equations with this node dynamics type. $f$ is represented by rhs field of OrdinaryNodeDynamics.

the general signature of rhs is

rhs(dint_dt::AbstractVector,
    u::Complex,
    i::Complex,
    int::AbstractVector,
    t,
    )::Complex

Input

u is the complex voltage $u$

i is the complex current $i$

int is the array of internal variables $y_1, \dots, y_n$

t is the time $t$

Output

the (complex) return value describes $\frac{du}{dt}$

rhs writes values in dint_dt describing the left-hand side $\frac{dy_1}{dt}, \dots, \frac{dy_n}{dt}$

The binary masses are:

m_u is the boolean value for $m_u$

m_int is the array of boolean values for $m^{int}_1, \dots, m^{int}_n$