Error Types

Error Types

PowerDynBase.PowerDynamicsError — Type.

Abstract super type of all PowerDynamics.jl Errors.

PowerDynBase.NodeDynamicsError — Type.

Error to be thrown if something goes wrong during the node dynamics construction.

PowerDynBase.GridDynamicsError — Type.

Error to be thrown if something goes wrong during the grid dynamics construction.

PowerDynBase.StateError — Type.

Error to be thrown if something goes wrong when creating or modifying states.

PowerDynSolve.GridSolutionError — Type.

Error to be thrown if something goes wrong during when solving a power grid model.