Node Types

Node Types

The currently implementes node types are

Purely Algebraic:

PowerDynBase.PQAlgebraic (PQ-bus)

PowerDynBase.PVAlgebraic (PV-bus)

PowerDynBase.SlackAlgebraic (Slack-bus / Vφ-bus)

Synchronous Machine Models:

PowerDynBase.SwingEq (2nd order)

PowerDynBase.SwingEqLVS (2nd order with an additional term for numerical voltage stability)

PowerDynBase.FourthEq (4th order)

Voltage Source Inverters:

PowerDynBase.VSIMinimal

PowerDynBase.VSIVoltagePT1

They are all subtypes of PowerDynBase.AbstractNodeParameters.

PowerDynBase.AbstractNodeParameters — Type.

Abstract super type for all node parameter types.

PowerDynBase.PQAlgebraic — Type.

PQAlgebraic(;S)

A node type that locally fixes the active ($P$) and reactive power ($Q$) output of the node.

Keyword Arguments

S = P + Q*im: the complex power output

Mathematical Representation

Using PQAlgebraic for node $a$ applies the equation

\[0 = S_a - u_a \cdot i_a^*.\]

source

PowerDynBase.PVAlgebraic — Type.

PVAlgebraic(;P,V)

A node type that locally fixes the active power ($P$) and the voltage magnitude ($V$) of the node.

Keyword Arguments

P: the active (real) power output

V: voltage magnitude

Mathematical Representation

Using PVAlgebraic for node $a$ applies the equations

\[0 = P_a - \Re\left(u_a \cdot i_a^*\right), \\ 0 = V_a - \left\|u_a\right\|.\]

source

PowerDynBase.SlackAlgebraic — Type.

SlackAlgebraic(;U)

A node type that locally fixes the complex voltage ($U$) of the node.

As the complex voltage can be represented as $U=Ve^{i\phi}$, this is equivlant to fixing the voltage magnitude $V$ and the angle $\phi$.

Keyword Arguments

U: the complex voltage

Mathematical Representation

Using SlackAlgebraic for node $a$ applies the equation

\[0 = U_a - u_a.\]

source

PowerDynBase.SwingEq — Type.

SwingEq(;H, P, D, Ω)

A node type that applies the swing equation to the frequency/angle dynamics and keeps the voltage magnitude as is.

Additionally to $u$, it has the internal dynamic variable $\omega$ representing the frequency of the rotator relative to the grid frequency $\Omega$, i.e. the real frequency $\omega_r$ of the rotator is given as $\omega_r = \Omega + \omega$.

Keyword Arguments

H: inertia

P: active (real) power output

D: damping coefficient

Ω: rated frequency of the power grid, often 50Hz

Mathematical Representation

Using SwingEq for node $a$ applies the equations

\[\frac{du_a}{dt} = i u_a \omega_a, \\ \frac{H}{2\pi\Omega}\frac{d\omega_a}{dt} = P_a - D_a\omega_a - \Re\left(u_a \cdot i_a^*\right),\]

which is equivalent to

\[\frac{d\phi_a}{dt} = \omega, \\ v = v(t=0) = \text{const.} \\ \frac{H}{2\pi\Omega}\frac{d\omega_a}{dt} = P_a - D_a\omega_a - \Re\left(u_a \cdot i_a^*\right),\]

source

PowerDynBase.SwingEqLVS — Type.

SwingEqLVS(;H, P, D, Ω, Γ, V)

A node type that applies the swing equation to the frequency/angle dynamics and has a linear voltage stability (LVS) term.

Additionally to $u$, it has the internal dynamic variable $\omega$ representing the frequency of the rotator relative to the grid frequency $\Omega$, i.e. the real frequency $\omega_r$ of the rotator is given as $\omega_r = \Omega + \omega$.

Keyword Arguments

H: inertia

P: active (real) power output

D: damping coefficient

Ω: rated frequency of the power grid, often 50Hz

Γ: voltage stability coefficient

V: set voltage, usually 1

Mathematical Representation

Using SwingEq for node $a$ applies the equations

\[\frac{du_a}{dt} = i u_a \omega - \frac{u}{\|u\|} Γ_a (v_a - V_a), \\ \frac{H}{2\pi\Omega}\frac{d\omega_a}{dt} = P_a - D_a\omega_a - \Re\left(u_a \cdot i_a^*\right),\]

which is equivalent to

\[\frac{d\phi_a}{dt} = \omega_a, \\ \frac{dv_a}{dt} = - Γ_a (v_a - V_a) \\ \frac{H}{2\pi\Omega}\frac{d\omega_a}{dt} = P_a - D_a\omega_a - \Re\left(u_a \cdot i_a^*\right),\]

source

PowerDynBase.FourthEq — Type.

FourthEq(H, P, D, Ω, E_f, T_d_dash ,T_q_dash ,X_q_dash ,X_d_dash,X_d, X_q)

A node type that applies the 4th-order synchronous machine model with frequency/angle and voltage dynamics.

Additionally to $u$, it has the internal dynamic variables

$\omega$ representing the frequency of the rotator relative to the grid frequency $\Omega$, i.e. the real frequency $\omega_r$ of the rotator is given as $\omega_r = \Omega + \omega$ and

$\theta$ representing the relative angle of the rotor with respect to the voltage angle $\phi$.

Keyword Arguments

H: inertia

P: active (real) power output

D: damping coefficient

Ω: rated frequency of the power grid, often 50Hz

T_d_dash: time constant of d-axis

T_q_dash: time constant of q-axis

X_d_dash: transient reactance of d-axis

X_q_dash: transient reactance of q-axis

X_d: reactance of d-axis

X_d: reactance of q-axis

Mathematical Representation

Using FourthEq for node $a$ applies the equations

\[ u = -je_c e^{j\theta} = -j(e_d + je_q)e^{j\theta}\\ e_c= e_d + je_q = jue^{-j\theta}\\ i = -ji'e^{j\theta} = -j(i_d+ j i_q )e^{j\theta} = Y^L \cdot u \\ i_c= i_d + ji_q = jie^{-j\theta}\\ p = \Re (i^* u)\]

The fourth-order equations read (according to Sauer, p. 140, eqs. (6110)-(6114)) and p. 35 eqs(3.90)-(3.91)

\[ \frac{d\theta}{dt} = \omega \\ \frac{d\omega}{dt} = P-D\omega - p -(x'_q-x'_d)i_d i_q\\ \frac{d e_q}{dt} = \frac{1}{T'_d} (- e_q - (x_d - x'_d) i_{d}+ e_f) \\ \frac{d e_d}{dt} = \frac{1}{T'_q} (- e_d + (x_q - x'_q) i_{q}) \\\]

With the PowerDynamics.jl \time{naming conventions} of $i$ and $u$ they read as

\[ \dot u = \frac{d}{dt}(-j e_c e^{j\theta})=-j(\dot e_d + j\dot e_q)e^{j\theta} + uj\omega\]

source

PowerDynBase.VSIMinimal — Type.

VSIMinimal(;τ_P,τ_Q,K_P,K_Q,E_r,P,Q)

A node type that applies the frequency and voltage droop control to control the frequency and voltage dynamics.

Additionally to $u$, it has the internal dynamic variable $\omega$ representing the frequency of the rotator relative to the grid frequency $\Omega$, i.e. the real frequency $\omega_r$ of the rotator is given as $\omega_r = \Omega + \omega$.

Keyword Arguments

τ_p: time constant active power measurement

τ_Q: time constant reactive power measurement

K_P: droop constant frequency droop

K_Q: droop constant voltage droop

V_r: reference/ desired voltage

P: active (real) power infeed

Q: reactive (imag) power infeed

Mathematical Representation

Using VSIMinimal for node $a$ applies the equations

\[\dot{\phi}_a=\omega_a\\ \dot{\omega}_a=\frac{1}{\tau_{P,a}}[-\omega_a-K_{P,a} (\Re\left(u_a \cdot i_a^*\right)-P_{ref,a})]\\ \tau_Q\dot{v}_a=-v_a+V_{ref}-K_{Q,a} (\Im\left(u_a \cdot i_a^*\right)-Q_{ref,a})\\ \dot{u}_a=\dot{v_a}e^{j\phi}+j\omega_a u_a\]

```

source

PowerDynBase.VSIVoltagePT1 — Type.

VSIVoltagePT1(;τ_v,τ_P,τ_Q,K_P,K_Q,E_r,P,Q)

A node type that applies the frequency and voltage droop control to control the frequency and voltage dynamics.

Additionally to $u$, it has the internal dynamic variable $\omega$ representing the frequency of the rotator relative to the grid frequency $\Omega$, i.e. the real frequency $\omega_r$ of the rotator is given as $\omega_r = \Omega + \omega$.

Keyword Arguments

τ_v: time constant voltage control delay

τ_p: time constant active power measurement

τ_Q: time constant reactive power measurement

K_P: droop constant frequency droop

K_Q: droop constant voltage droop

V_r: reference/ desired voltage

P: active (real) power infeed

Q: reactive (imag) power infeed

Mathematical Representation

Using VSIVoltagePT1 for node $a$ applies the equations

\[\dot{\phi}_a=\omega_a\\ \dot{\omega}_a=\frac{1}{\tau_{P,a}}[-\omega_a-K_{P,a} (\Re\left(u_a \cdot i_a^*\right)-P_{ref,a})]\\ \tau_v\dot{v}_{a}=-v_a+V_{ref}-K_{Q,a}(q_{m,a}-Q_{ref,a})\\ \tau_Q \dot{q}_{m,a}=-q_{m,a}+\Im\left(u_a \cdot i_a^*\right)\\ \dot{u}_a=\dot{v_a}e^{j\phi}+j\omega_a u_a\\\]

source