Error Types

Error Types

PowerDynamics.PowerDynamicsError — Type.

Abstract super type of all PowerDynamics.jl Errors.

source

PowerDynamics.NodeDynamicsError — Type.

Error to be thrown if something goes wrong during the node dynamics construction.

source

PowerDynamics.StateError — Type.

Error to be thrown if something goes wrong when creating or modifying states.

source

PowerDynamics.GridSolutionError — Type.

Error to be thrown if something goes wrong during power grid solving

source

PowerDynamics.OperationPointError — Type.

Error to be thrown if something goes wrong during the operation point search.

source