Error Types

Error Types

PowerDynamics.PowerDynamicsError — Type.

Abstract super type of all PowerDynamics.jl Errors.

source

PowerDynamics.NodeDynamicsError — Type.

Error to be thrown if something goes wrong during the node dynamics construction.

source

Missing docstring.

Missing docstring for PowerDynamics.GridDynamicsError. Check Documenter's build log for details.

PowerDynamics.StateError — Type.

Error to be thrown if something goes wrong when creating or modifying states.

source

Missing docstring.

Missing docstring for PowerDynamics.GridSolutionError. Check Documenter's build log for details.