Line Types

Line Types

The currently implemented line types are

PiModelLine

RLLine

StaticLine

Transformer

Detailed Line Type Documentation

PowerDynamics.PiModelLine — Type.

    PiModelLine(from, to, y, y_shunt_km, y_shunt_mk)

A line modelled according to the PI-Model.

See also the Chapter 2 in Göran Andersson, Power System Analysis, Lecture 227-0526-00, ITET ETH Zurich, 2012

Arguments

from : node k

to : node m

y: admittance of line between k and m

y_shunt_km: shunt admittance at the end connected to node k

y_shunt_mk: shunt admittance at the end connected to node m

t_km: transformer ratio at the end connected to node k

t_mk: transformer ratio at the end connected to node m

Assumptions:

the line admittance is symmetric

source

PowerDynamics.RLLine — Type.

RLLine(; from, to, R, L, ω0)

dynamic line with series resistance R and series inductance L.

Keyword Arguments

from: start node of the line

to: end node of the line

R: series resistance R

L: series inductance L

ω0: rated frequency in [rad/s] of the power grid, often $2π50$Hz

Mathematical Representation

Using RLLine for the line from–to applies Eqn. (2) from Brouillon, J. S., Colombino, M., Groß, D., & Dörfler, F. (2018). The effect of transmission-line dynamics on a globally synchronizing controller for power inverters. In 2018 European Control Conference (ECC) (pp. 2242-2247). IEEE.

source

PowerDynamics.StaticLine — Type.

    StaticLine(from, to, Y)

A static model that represents a line with an admittance Y.

Arguments

from : start node

to : end node

Y: scalar admittance of line between from and to

source

PowerDynamics.Transformer — Type.

    Transformer(from, to, y, t_ratio)

assuming an ideal transformer in series with an admittance. The representation uses the Π model.

Mathematical Representation

The voltage transforms as:

\[ u_{to} = t_{ratio} u_{from}\]

Arguments

from : start node

to : end node

y: transformer admittance

t_ratio: transformation ration

Assumptions

The admittance is here taken to be on the high-voltage side.

source

PowerDynamics.AbstractLine — Type.

Abstract base type for all lines

source