Line Types

Line Types

The currently implemented line types are

PiModelLine

StaticLine

Transformer

Detailed Line Type Documentation

PowerDynamics.PiModelLine — Type.

PiModelLine(from, to, y, y_shunt_km, y_shunt_mk)

A line modelled according to the PI-Model.

source

PowerDynamics.StaticLine — Type.

StaticLine(from, to, Y)

A static model that represents a line with an admittance Y.

source

PowerDynamics.Transformer — Type.

Transformer(from, to, y, t_ratio)

A transformer representation uses the Π model, assuming an ideal transformer in series with an admittance. The admittance is here taken to be on the high-voltage side.

source

PowerDynamics.AbstractLine — Type.

Abstract base type for all lines

source