Node Dynamics Types

Node Dynamics Types

In this section, the implemented general types of node dynamics with the corresponding helper functions and constants are introduced. The documentation is done for each type below. The main types are:

PowerDynBase.OrdinaryNodeDynamics

PowerDynBase.OrdinaryNodeDynamicsWithMass

Documentation

PowerDynBase.no_internal_differentials — Constant.

A variable to be used when no internal differentials are present for a node dynamics type.

PowerDynBase.no_internal_masses — Constant.

A variable to be used when no internal masses are present for a node dynamics type.

PowerDynBase.OrdinaryNodeDynamicsWithMass — Type.

OrdinaryNodeDynamicsWithMass(;rhs, n_int, m_u, m_int)

The type representing the dynamics of a node that is described via ODEs.

Each node $a$ has the complex voltage $u$ and $n$ (= n_int) real internal variables $y_1, \dots, y_n$, so it generally describes a system of ordinary differential equation with a voltage mass $m_u$ and internal masses $m^{int}_1, \dots, m^{int}_n$ as

\[m_u\frac{du_a}{dt} = f_u(u_a, {i_c}_a, y_1, \dots, y_n) \\ m^{int}_k\frac{dy_{ak}}{dt} = f_k(u_a, {i_c}_a, y_1, \dots, y_n)\quad \forall k = 1, \dots, n.\]

As we assume that all masses are binary (either 1, or 0), that means, one can implement semi-explicit differential algebraic equations with this node dynamics type.

No documentation found.

Markdown.@doc_str is a macro.

# 1 method for macro "@doc_str":
[1] @doc_str(__source__::LineNumberNode, __module__::Module, s::AbstractString, t...) in Markdown at /buildworker/worker/package_linux64/build/usr/share/julia/stdlib/v1.0/Markdown/src/Markdown.jl:53

The binary masses are:

m_u is the boolean value for $m_u$

m_int is the array of boolean values for $m^{int}_1, \dots, m^{int}_n$

Base.convert — Method.

convert(::Type{OrdinaryNodeDynamicsWithMass}, ::OrdinaryNodeDynamics)

Conversion of OrdinaryNodeDynamics to OrdinaryNodeDynamicsWithMass by assuming all masses are 1 (true).

Base.convert — Method.

convert(::Type{AlgebraicNodeDynamics}, ::OrdinaryNodeDynamicsWithMass)

Conversion of OrdinaryNodeDynamicsWithMass to AlgebraicNodeDynamics by transforming a RHS function into a root function.

Base.convert — Method.

convert(::Type{AlgebraicNodeDynamics}, ::OrdinaryNodeDynamics)

Conversion of OrdinaryNodeDynamics to AlgebraicNodeDynamics by going via OrdinaryNodeDynamicsWithMass.

Base.promote_rule — Method.

promote_rule(::Type{OrdinaryNodeDynamics}, ::Type{OrdinaryNodeDynamicsWithMass}) = OrdinaryNodeDynamicsWithMass

OrdinaryNodeDynamics can be promoted to OrdinaryNodeDynamicsWithMass, see DPSABase.convert.

Base.promote_rule — Method.

promote_rule(::Type{OrdinaryNodeDynamicsWithMass}, ::Type{AlgebraicNodeDynamics}) = AlgebraicNodeDynamics

OrdinaryNodeDynamicsWithMass can be promoted to AlgebraicNodeDynamics, see DPSABase.convert.

Base.promote_rule — Method.

promote_rule(::Type{OrdinaryNodeDynamics}, ::Type{AlgebraicNodeDynamics}) = AlgebraicNodeDynamics

OrdinaryNodeDynamics can be promoted to AlgebraicNodeDynamics, see DPSABase.convert.