Node Types

Node Types

The currently implementes node types are

Detailed Node Type Documentation

Missing docstring.

Missing docstring for PowerDynamics.AbstractNodeParameters. Check Documenter's build log for details.

PowerDynamics.CSIMinimal — Type.

CSIMinimal(;I_r)

A node type that keeps the current fixed as a desired values I_r.

CSIMinimal models an inverters as an ideal current source. This can be the most simple representation of an inverter in grid-feeding mode, according to Rocabert, Joan, et al. "Control of power converters in AC microgrids." (2012). Here, additionally to u, there are no internal dynamic variables.

Keyword Arguments

I_r: reference/ desired current

Mathematical Representation

Using CSIMinimal for node $a$ gives:

\[0 = I_{r,a} - \left\|i_a\right\|\]

source

PowerDynamics.ExponentialRecoveryLoad — Type.

ExponentialRecoveryLoad(P0, Q0, Nps, Npt, Nqs, Nqt, Tp, Tq, V0)

A node type that represents the exponential recovery load model. The exponential recovery load model aims to capture the load restoration characteristics with an exponential recovery process expressed as an input–output relationship between powers (real and reactive) and voltage.

Keyword Arguments

P0: Active power load demand [pu]

Q0: Reactive power load demand [pu]

Nps: Steady-state load voltage dependence p-axis [pu]

Npt: Transient load voltage dependence p-axis [pu]

Nqs: Steady-state load voltage dependence q-axis [pu]

Nqt: Transient load voltage dependence q-axis [ pu]

Tp: Load recovery constant p-axis [s]

Tq: Load recovery constant q-axis [s]

V0: Reference grid voltage [pu]

Mathematical Representation

\[ \dfrac{dx_p}{dt} = \dfrac{1}{T_p}(-x_p + P_0(\dfrac{abs(u)}{V_0})^{N_{ps}} - P_0(\dfrac{abs(u)}{V_0})^{N_{pt}}) \\ \dfrac{dx_q}{dt} = \dfrac{1}{T_q}(-x_q + Q_0(\dfrac{abs(u)}{V_0})^{N_{qs}} - Q_0(\dfrac{abs(u)}{V_0})^{N_{qt}}) \\ P = x_p + P_0(\dfrac{abs(u)}{V_0})^{N_{pt}} \\ Q = x_q - Q_0(\dfrac{abs(u)}{V_0})^{N_{qt}} \\\]

IEEE TRANSACTIONS ON POWER SYSTEMS, VOL. 21, NO. 3, AUGUST 2006 Measurement-Based Dynamic Load Models: Derivation, Comparison, and Validation Byoung-Kon Choi, Member, IEEE, Hsiao-Dong Chiang, Fellow, IEEE, Yinhong Li, Hua Li, Member, IEEE, Yung-Tien Chen, Der-Hua Huang, and Mark G. Lauby

source

PowerDynamics.FourthOrderEq — Type.

FourthEq(H, P, D, Ω, E_f, T_d_dash ,T_q_dash ,X_q_dash ,X_d_dash,X_d, X_q)

A node type that applies the 4th-order synchronous machine model with frequency/angle and voltage dynamics, which is implemented according to P. Sauer, "Power System Dynamics and Stability". For an illustration of a synchronous machine schematic see P. Sauer, Fig. 3.1 on p. 25.

Usually the swing equation (SwingEq) is used for short time periods to analyze the transient behavior of generators in a power grid, the so-called first swing. The 4th-order model also takes the back reaction of the power flow onto the voltage into account. This has the effect that the angle of the voltage as seen by the power grid, and the angle of the rotating mass are no longer the same but become dynamically coupled.

Additionally to $u$, it has the internal dynamic variables

$ω$ representing the frequency of the rotator relative to the grid frequency $Ω$, i.e. the real frequency $ω_r$ of the rotator is given as $\omega_r = \Omega + \omega$ and

$θ$ representing the relative angle of the rotor with respect to the voltage angle $ϕ$.

Keyword Arguments

H: shaft inertia constant (given in [s]), defined according to P. Sauer, p. 33, eq. (3.60)

P: active (real) power output, also called the mechanical torque applied to the shaft, given in [pu]

D: damping coefficient (given in [s], see P. Sauer, eq. (5.156) where the damping torque is equal Dω)

Ω: rated frequency of the power grid, often 50Hz

T_d_dash: time constant of d-axis, given in [s], see P. Sauer, chapter 3.7, p. 54 for a general explanation on time constants

T_q_dash: time constant of q-axis, given in [s]

X_d_dash: transient reactance of d-axis, given in [pu]

X_q_dash: transient reactance of q-axis, given in [pu]

X_d: reactance of d-, given in [pu]

X_d: reactance of q-axis, given in [pu]

E_f: scaled field voltage, which, if set equal to 1.0 pu, gives 1.0 pu open-circuit terminal voltage. The physical device that provides the value of E_f is called the exciter (according to P. Sauer, p. 65)

Mathematical Representation

Using FourthEq for node $a$ applies the equations

\[ u = -je_c e^{j\theta} = -j(e_d + je_q)e^{j\theta}\\ e_c= e_d + je_q = jue^{-j\theta}\\ i = -ji'e^{j\theta} = -j(i_d+ j i_q )e^{j\theta} = Y^L \cdot u \\ i_c= i_d + ji_q = jie^{-j\theta}\\ p = \Re (i^* u)\]

where complex voltage and current are described in a co-rotating frame with axes labeled d and q.

The fourth-order equations read (according to P. Sauer, "Power System Dynamics and Stability", p. 140, eqs. (6110)-(6114)) and p. 35 eqs(3.90)-(3.91)

\[ \frac{d\theta}{dt} = \omega \\ \frac{d\omega}{dt} = (P-D\omega - p -(X'_q-X'_d)i_d i_q)Ω_H\\ \frac{d e_q}{dt} = \frac{1}{T'_d} (- e_q - (X_d - X'_d) i_{d}+ E_f) \\ \frac{d e_d}{dt} = \frac{1}{T'_q} (- e_d + (X_q - X'_q) i_{q}) \\\]

The equations for frequency and phase represent energy conservation and phase shift. The dynamic equations for the complex voltage show the relationship between the dynamicy of flux linkages and currents which must reflect a conservative coupling field.

With the PowerDynamics.jl naming conventions of $i$ and $u$ they read as

\[ \dot u = \frac{d}{dt}(-j e_c e^{j\theta})=-j(\dot e_d + j\dot e_q)e^{j\theta} + uj\omega\]

source

PowerDynamics.FourthOrderEqGovernorExciterAVR — Type.

FourthOrderEqGovernorExciterAVR(H, P, D, Ω, T_d_dash ,T_q_dash ,X_q_dash ,X_d_dash,X_d, X_q, T_e, T_a, T_f, K_e, K_a, K_f, V_ref, R_d, T_sv, T_ch)

A node type that applies the 4th-order synchronous machine model with frequency/angle and voltage dynamics, including an Exciter, Automatic Voltage Regulator and Governor which is implemented according to P. Sauer, "Power System Dynamics and Stability". For an illustration of a synchronous machine schematic see P. Sauer, Fig. 3.1 on p. 25.

Exciter and Automatic Voltage Regulator: The equations for the systems that balance the AC synchronous machine voltage level by increasing or decreasing the exciter DC voltage. Note, within this model, the transient reactance in the d-axis of the generator needs to be included into the nodal admittance matrix. As the bus of this generator node type is constructed to be an internal generator bus.

Governor: The prime mover provides the mechanism for controlling the synchronous machine speed and, hence, terminal voltage frequency.

Usually the swing equation (SwingEq) is used for short time periods to analyze the transient behavior of generators in a power grid, the so-called first swing. The 4th-order model also takes the back reaction of the power flow onto the voltage into account. This has the effect that the angle of the voltage as seen by the power grid, and the angle of the rotating mass are no longer the same but become dynamically coupled.

Additionally to $u$, it has the internal dynamic variables

$ω$ representing the frequency of the rotator relative to the grid frequency $Ω$, i.e. the real frequency $ω_r$ of the rotator is given as $\omega_r = \Omega + \omega$ and

$θ$ representing the relative angle of the rotor with respect to the voltage angle $ϕ$.

Keyword Arguments

H: shaft inertia constant (given in [s]), defined according to P. Sauer, p. 33, eq. (3.60)

P: active (real) power output, also called the mechanical torque applied to the shaft, given in [pu]

D: damping coefficient (given in [s], see P. Sauer, eq. (5.156) where the damping torque is equal Dω)

Ω: rated frequency of the power grid, often 50Hz

T_d_dash: time constant of d-axis, given in [s], see P. Sauer, chapter 3.7, p. 54 for a general explanation on time constants

T_q_dash: time constant of q-axis, given in [s]

X_d_dash: transient reactance of d-axis, given in [pu]

X_q_dash: transient reactance of q-axis, given in [pu]

X_d: reactance of d-, given in [pu]

X_d: reactance of q-axis, given in [pu]

'T_e' : Exciter time constant, integration rate associated with exciter control [s]

'T_a' : Maximum voltage regulator output [s]

'T_f' : Excitation control system stabilizer time constant [s]

'K_e' : Exciter constant related to self-excited field [pu]

'K_a' : Voltage Regulator gain [pu]

'K_f' : Excitation control system stabilizer gains [pu]

'V_ref' : Reference voltage for the AVR [pu]

'Rd' : Speed regulation [2πdroop/ωs]

'T_sv' : Steam Valve time constant [s]

'T_ch' : Steam Chest time constant [s]

Mathematical Representation Synchronous Machine

Using FourthEq for node $a$ applies the equations

\[ u = -je_c e^{j\theta} = -j(e_d + je_q)e^{j\theta}\\ e_c= e_d + je_q = jue^{-j\theta}\\ i = -ji'e^{j\theta} = -j(i_d+ j i_q )e^{j\theta} = Y^L \cdot u \\ i_c= i_d + ji_q = jie^{-j\theta}\\ p = \Re (i^* u) \\\]

where complex voltage and current are described in a co-rotating frame with axes labeled d and q.

The fourth-order equations read (according to P. Sauer, "Power System Dynamics and Stability", p. 140, eqs. (6110)-(6114)) and p. 35 eqs(3.90)-(3.91)

\[ \frac{d\theta}{dt} = \omega \\ \frac{d\omega}{dt} = (P-D\omega - p -(X'_q-X'_d)i_d i_q)Ω_H\\ \frac{d e_q}{dt} = \frac{1}{T'_d} (- e_q - (X_d - X'_d) i_{d}+ E_f) \\ \frac{d e_d}{dt} = \frac{1}{T'_q} (- e_d + (X_q - X'_q) i_{q}) \\\]

Exciter and AVR equations

\[ u_{terminal} = e'_c - j X'_d i \\ S_{e}(e_{fd}) = 0.098e^{0.55 e_{fd}} (according to P. Sauer, p. 70) \\ \dfrac{dR_f}{dt} = \dfrac{1}{T_f} (-R_f + \dfrac{K_f}{T_f} e_f) \\ \dfrac{dv_r}{dt} = \dfrac{1}{T_a} (-v_r + (K_a R_f) -\dfrac{K_a K_f}{T_f}e_{fd} + K_a (V_{ref} - abs(u_{terminal}))) \\ \dfrac{de_{fd}}{dt} = \dfrac{1}{T_e} (-K_e + S_{e}(e_{fd})e_{fd} + v_r) \\\]

Governor equations

\[ \dfrac{dP_m}{dt} = \dfrac{1}{T_{ch}} (-P_m + P_{sv}) \\ Assumption: T_m = P_m \\ \dfrac{dP_{sv}}{dt} = \dfrac{1}{T_{sv}} (-P_{sv} + P_c -\dfrac{1}{R_d} (\dfrac{\omega}{\omega_s} - 1)) \\\]

The equations for frequency and phase represent energy conservation and phase shift. The dynamic equations for the complex voltage show the relationship between the dynamicy of flux linkages and currents which must reflect a conservative coupling field.

With the PowerDynamics.jl naming conventions of $i$ and $u$ they read as

\[ \dot u = \frac{d}{dt}(-j e_c e^{j\theta})=-j(\dot e_d + j\dot e_q)e^{j\theta} + uj\omega \\\]

source

PowerDynamics.PQAlgebraic — Type.

PQAlgebraic(;S)

A node type that locally fixes the active ($P$) and reactive power ($Q$) output of the node.

Keyword Arguments

S = P + Q*im: the complex power output

Mathematical Representation

Using PQAlgebraic for node $a$ applies the equation

\[0 = S_a - u_a \cdot i_a^*.\]

source

PowerDynamics.PVAlgebraic — Type.

PVAlgebraic(;P,V)

A node type that locally fixes the active power ($P$) and the voltage magnitude ($V$) of the node.

Keyword Arguments

P: the active (real) power output

V: voltage magnitude

Mathematical Representation

Using PVAlgebraic for node $a$ applies the equations

\[0 = P_a - \Re\left(u_a \cdot i_a^*\right), \\ 0 = V_a - \left\|u_a\right\|.\]

source

PowerDynamics.PowerGrid — Type.

Encapsulates nodes & lines of the power grid and a graph connecting both.

source

PowerDynamics.SlackAlgebraic — Type.

SlackAlgebraic(;U)

A node type that locally fixes the complex voltage ($U$) of the node.

As the complex voltage can be represented as $U=Ve^{i\phi}$, this is equivlant to fixing the voltage magnitude $V$ and the angle $\phi$.

Keyword Arguments

U: the complex voltage

Mathematical Representation

Using SlackAlgebraic for node $a$ applies the equation

\[0 = U_a - u_a.\]

source

PowerDynamics.SwingEq — Type.

SwingEq(;H, P, D, Ω)

A node type that applies the swing equation to the frequency/angle dynamics and keeps the voltage magnitude as is.

Additionally to $u$, it has the internal dynamic variable $\omega$ representing the frequency of the rotator relative to the grid frequency $\Omega$, i.e. the real frequency $\omega_r$ of the rotator is given as $\omega_r = \Omega + \omega$.

Keyword Arguments

H: inertia

P: active (real) power output

D: damping coefficient

Ω: rated frequency of the power grid, often 50Hz

Mathematical Representation

Using SwingEq for node $a$ applies the equations

\[\frac{du_a}{dt} = i u_a \omega_a, \\ \frac{H}{2\pi\Omega}\frac{d\omega_a}{dt} = P_a - D_a\omega_a - \Re\left(u_a \cdot i_a^*\right),\]

which is equivalent to

\[\frac{d\phi_a}{dt} = \omega, \\ v = v(t=0) = \text{const.} \\ \frac{H}{2\pi\Omega}\frac{d\omega_a}{dt} = P_a - D_a\omega_a - \Re\left(u_a \cdot i_a^*\right),\]

source

PowerDynamics.SwingEqLVS — Type.

SwingEqLVS(;H, P, D, Ω, Γ, V)

A node type that applies the swing equation to the frequency/angle dynamics and has a linear voltage stability (LVS) term.

Additionally to $u$, it has the internal dynamic variable $\omega$ representing the frequency of the rotator relative to the grid frequency $\Omega$, i.e. the real frequency $\omega_r$ of the rotator is given as $\omega_r = \Omega + \omega$.

Keyword Arguments

H: inertia

P: active (real) power output

D: damping coefficient

Ω: rated frequency of the power grid, often 50Hz

Γ: voltage stability coefficient

V: set voltage, usually 1

Mathematical Representation

Using SwingEq for node $a$ applies the equations

\[\frac{du_a}{dt} = i u_a \omega - \frac{u}{\|u\|} Γ_a (v_a - V_a), \\ \frac{H}{2\pi\Omega}\frac{d\omega_a}{dt} = P_a - D_a\omega_a - \Re\left(u_a \cdot i_a^*\right),\]

which is equivalent to

\[\frac{d\phi_a}{dt} = \omega_a, \\ \frac{dv_a}{dt} = - Γ_a (v_a - V_a) \\ \frac{H}{2\pi\Omega}\frac{d\omega_a}{dt} = P_a - D_a\omega_a - \Re\left(u_a \cdot i_a^*\right),\]

source

PowerDynamics.VSIMinimal — Type.

VSIMinimal(;τ_P,τ_Q,K_P,K_Q,E_r,P,Q)

A node type that applies the frequency and voltage droop control to control the frequency and voltage dynamics.

VSIMinimal models an inverters as AC voltage source which means the amplitude and frequency can defined by the designer (often called grid-forming inverter mode). The frequency and voltage regulation is assumed to be instantaneous. In addition simple proportional controllers are implemented for frequency and voltage such that the frequency ω and voltage amplitudes v of the inverters are modified depending on the deviations (with respect to a desired value) of the active and reactive powers, respectively. ift is assumed that active and reactive power are measured via low pass fileters with time constant τ_P and τ_Q, respectively. VSIMinimal can be derived from VSIVoltagePT1 by assuming an instantaneous voltage regulation without delay.

Additionally to $u$, it has the internal dynamic variable $ω$ representing the frequency of the rotator relative to the grid frequency $Ω$, i.e. the real frequency $ω_r$ of the rotator is given as $ω_r = Ω + ω$.

Keyword Arguments

τ_p: time constant active power measurement

τ_Q: time constant reactive power measurement

K_P: droop constant frequency droop

K_Q: droop constant voltage droop

V_r: reference/ desired voltage

P: active (real) power infeed

Q: reactive (imag) power infeed

Mathematical Representation

Using VSIMinimal for node $a$ (according to J. Schiffer et. al., eq. (7)) gives the equations

\[\dot{\phi}_a=\omega_a\\ \dot{\omega}_a=\frac{1}{\tau_{P,a}}[-\omega_a-K_{P,a} (\Re\left(u_a \cdot i_a^*\right)-P_{a})]\\ \tau_Q\dot{v}_a=-v_a+V_{r}-K_{Q,a} (\Im\left(u_a \cdot i_a^*\right)-Q_{a})\\ \dot{u}_a=\dot{v_a}e^{j\phi}+j\omega_a u_a\]

```

source

PowerDynamics.VSIVoltagePT1 — Type.

VSIVoltagePT1(;τ_v,τ_P,τ_Q,K_P,K_Q,E_r,P,Q)

A node type that applies the frequency and voltage droop control to control the frequency and voltage dynamics.

VSIVoltagePT1 models an inverters as AC voltage source which means the amplitude and frequency can defined by the designer (often called grid-forming inverter mode). The frequency regulation is assumed to be instantaneous, but the voltage control happens with a delay τ_v that is represented by a first order filter. In addition simple proportional controllers are implemented for frequency and voltage such that the frequency ω and voltage amplitudes v of the inverters are modified depending on the deviations (with respect to a desired value) of the active and reactive powers, respectively. ift is assumed that active and reactive power are measured via low pass fileters with time constant τ_P and τ_Q, respectively.

Hence, additionally to u, it has the internal dynamic variables

ω representing the frequency of the rotator relative to the grid frequency $Ω$, i.e. the real frequency $ω_r$ of the rotator is given as $ω_r = Ω + ω$.

q_m is the measured reactive power at the grid connection point.

Keyword Arguments

τ_v: time constant voltage control delay

τ_P: time constant active power measurement

τ_Q: time constant reactive power measurement

K_P: droop constant frequency droop

K_Q: droop constant voltage droop

V_r: reference/ desired voltage

P: active (real) power infeed

Q: reactive (imag) power infeed

Mathematical Representation

Using VSIVoltagePT1 for node $a$ (according to J. Schiffer et. al., eq. (6)) gives the equations

\[\dot{\phi}_a=\omega_a\\ \dot{\omega}_a=\frac{1}{\tau_{P,a}}[-\omega_a-K_{P,a} (\Re\left(u_a \cdot i_a^*\right)-P_{ref,a})]\\ \tau_v\dot{v}_{a}=-v_a+V_{ref}-K_{Q,a}(q_{m,a}-Q_{ref,a})\\ \tau_Q \dot{q}_{m,a}=-q_{m,a}+\Im\left(u_a \cdot i_a^*\right)\\ \dot{u}_a=\dot{v_a}e^{j\phi}+j\omega_a u_a\\\]

In general $τ_V ≪ τ_P$, assuming $τ_V = 0$ would then lead to VSIMinimal.

source

PowerDynamics.OperationPointError — Type.

Error to be thrown if something goes wrong during the operation point search.

source

PowerDynamics.PowerGridSolution — Type.

struct PowerGridSolution
    dqsol::AbstractTimeseriesSolution
    powergrid::PowerGrid
end

The data structure interfacing to the solution of the differntial equations of a power grid. Normally, it is not created by hand but return from PowerDynSolve.solve.

Accessing the solution in a similar interface as State.

For some grid solution sol, one can access the variables as

sol(t, n, s)

where t is the time (either float or array), n the node number(s) (either integer, array, range (e.g. 2:3) or colon (:, for all nodes)), and s is the symbol represnting the chosen value. s can be either: :v, :φ, :i, :iabs, :δ, :s, :p, :q, or the symbol of the internal variable of the nodes. The meaning of the symbols derives from the conventions of PowerDynamics.jl. Finally, one can access the a-th internal variable of a node by using sol(t, n, :int, a).

Interfacing the Plots.jl library via plotting recipes, that follow similar instructions as the direct access to the solution.

For some grid solution sol, one plot variables of the solution asin

using Plots
plot(sol, n, s, plots_kwargs...)

where n and s are as in the accessing of plots, and plots_kwargs are the keyword arguments for Plots.jl.

source