PowerGrid model

PowerGrid model

PowerDynamics.systemsize — Method.

Returns the total size of dynamical variables of the whole powergrid

source

PowerDynamics.PowerGrid — Type.

Powergrid(graph, nodes, lines)

Our model describes a powergrid as an undirected graph consisting of edges that represent electrical lines and vertices that represent specific electrical nodes i.e. generators, inverters etc.

source

PowerDynamics.PowerGrid — Method.

Powergrid(nodes, lines)

creates a PowerGrid from nodes and lines. The underlying graph is created automatically.

source