Simulations or Fault Scenarios

Simulations or Fault Scenarios

After having defined a PowerGrid a simulation can be run.

Currently PowerDynamics supports the following simulations:

ChangeInitialConditions

LineFailure

NodeParameterChange

NodeShortCircuit

PowerPerturbation

Detailed Simulation/ Fault Type Documentation

PowerDynamics.NodeShortCircuit — Type.

    NodeShortCircuit(;node,Y,tspan_fault)

Keyword Arguments

node: number or name of the node

Y: admittance of the short circuit

tspan_fault: short circuit timespan

Optional Keyword Arguments

var: symbol of the shunt admittance, default: :Y_n

source

PowerDynamics.NodeParameterChange — Type.

NodeParameterChange(;node, value, tspan_fault, var)

Keyword Arguments

node: number or name of the node

value: value of power during fault event

tspan_fault: PowerPerturbation timespan

var: symbol of parameter that is perturbed

source

PowerDynamics.PowerPerturbation — Type.

PowerPerturbation(;node_number,fault_power,tspan_fault,var)

Keyword Arguments

node_number: number or name of the node

fault_power: value of power during fault event

tspan_fault: PowerPerturbation timespan

var: parameter symbol on the node that represents Power, default is :P

source

PowerDynamics.ChangeInitialConditions — Type.

ChangeInitialConditions(;node,var,f)

Keyword Arguments

node: number of the node

var: symbol of the variable to be perturbated

f: function for mapping the variable x to the perturbated value

source

PowerDynamics.simulate — Method.

simulate(p::ChangeInitialConditions, powergrid, x0; timespan)

Simulates a ChangeInitialConditions

source

PowerDynamics.LineFailure — Type.

LineFailure(;line_name,tspan_fault)

The arguments line_name and tspan_fault specify the line that should be disconnected from the grid and the respective fault duration. For a list of lines the line_name is the index and for an OrderedDict it is the key of the line.

source

PowerDynamics.LineFault — Type.

LineFault(;from,to)

The arguments from and to specify the line that should be disconnected from the grid.

source

PowerDynamics.simulate — Method.

simulate(lf::LineFault, powergrid, x0; timespan)

Simulates a LineFault

source

PowerDynamics.AbstractPerturbation — Type.

AbstractPerturbation(;)

An AbstractPerturbation can be uses to construct almost any type of perturbation that has a time span of a fault. Perturbations that are of type of AbstractPerturbation can use the same simulate function then. For example implementations of faults as type AbstractPerturbation see e.g. PowerPerturbation.

Keyword Arguments

tspan_fault: number of the node

source

PowerDynamics.FieldUpdateError — Type.

Error to be thrown if something goes wrong during power perturbation

source

PowerDynamics.simulate — Method.

simulate(no::AbstractPerturbation, powergrid, x1, timespan)

Simulates a AbstractPerturbation

source