Simulations

Simulations

After having defined a PowerGrid a simulation can be run.

Currently PowerDynamics supports the following simulations:

Perturbation

LineFault

NodeShortCircuit

PowerPerturbation

PowerDynamics.simulate — Method.

simulate(lf::LineFault, powergrid, x0; timespan)

Simulates a LineFault

source

PowerDynamics.simulate — Method.

simulate(p::Perturbation, powergrid, x0; timespan)

Simulates a Perturbation

source

PowerDynamics.LineFault — Type.

LineFault(;from,to)

The arguments from and to specify the line that should be disconnected from the grid.

source

PowerDynamics.Perturbation — Type.

Perturbation(;node,var,f)

Keyword Arguments

node: number of the node

var: symbol of the variable to be perturbated

f: function for mapping the variable x to the perturbated value

source

PowerDynamics.simulate — Method.

simulate(nsc::NodeShortCircuit, powergrid, x1, timespan)

Simulates a NodeShortCircuit

source

PowerDynamics.NodeShortCircuit — Type.

NodeShortCircuit(;node,var,f)

Keyword Arguments

node: number of the node

R: resistance of the shortcircuit

sc_timespan: shortcircuit timespan

source

PowerDynamics.simulate — Method.

simulate(nsc::PowerPerturbation, powergrid, x0, timespan)

Simulates a PowerPerturbation

source

PowerDynamics.PowerPerturbation — Type.

PowerPerturbation(;node_number,fraction,tspan_fault)

Keyword Arguments

node_number: number of the node

fraction: percentage factor to be applied to the active power P

tspan_fault: PowerPerturbation timespan

source